restructure schema and types
