+++ src/lib/data-utils/domains.ts
@@ -2,8 +2,8 @@ import type { Database } from 'sqlite';
 
 import { domain } from '../models/domain';
 import type { Domain } from '../models/domain';
-import type { BrowseParamsSchema } from '../types/params';
-import type { BrowseQueryStringSchema } from '../types/querystring';
+import type { BrowseParamsSchema } from '../types/schemas';
+import type { BrowseQueryStringSchema } from '../types/schemas';
 
 const BROWSE_FILTER_MAP: Record<
   BrowseParamsSchema['filter'],
