--- src/lib/types/params.d.ts
@@ -1,10 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface BrowseParamsSchema {
-  filter: 'all' | 'numbers' | 'a_e' | 'f_j' | 'k_o' | 'p_t' | 'u_z';
-}
