--- src/lib/types/querystring.d.ts
@@ -1,11 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface BrowseQueryStringSchema {
-  page: number;
-  size: 10 | 20 | 50 | 100;
-}
